ngo_project/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
folium>=0.14.0
plotly>=5.13.0
pyperclip>=1.8.2
